requirements.txt
anvil==0.0.2
anvil_uplink==0.4.0
numpy==1.23.1
onnx==1.12.0
onnxruntime==1.12.1
pandas==1.4.3
parlai==1.7.1
projects==1.0.9
tensorflow==2.10.0
tf2onnx==1.12.1
tf_models_official==2.5.0
tqdm==4.64.0
transformers==4.21.1
git+https://github.com/LuCeHe/GenericTools